requirements.txt
requests
ratelimit
Pillow
urllib3